requirements.txt
boto3
pandas
requests
awswrangler
python-dotenv